Extract Transit Gateway Options
Extracts the transit gateway options to their own complex type to replicate the api structure.
